refactor: P2-G Step2 迁移存档逻辑到 persistence（薄委托）
- 新建 src/engine/persistence.py：save/load/recover_all/import_save_zip/save_all_active
  + _save_path/_chatlog_path/_append_chatlog/_truncate_chatlog/_restore_chatlog 迁出
- GameRegistry 保留薄委托（公共方法 + 被外部调用的 _save_path）；
  纯私有 4 个（chatlog 相关）无外部调用不设委托
- 常量动态引用 game_instance.MAX_SAVE_*（模块引用），保持测试 monkeypatch 生效
- 清理 game_instance 顶部 unused import（io/zipfile/time）
- 739 测试通过，ruff 全过
